+++ q_compress/src/gcd_utils.rs
@@ -1,5 +1,6 @@
 use crate::data_types::{NumberLike, UnsignedLike};
-use crate::{BitReader, BitWriter, Flags, Prefix};
+use crate::bit_reader::BitReader;
+use crate::{BitWriter, Flags, Prefix};
 use crate::errors::{QCompressError, QCompressResult};
 
 // fast if b is small, requires b > 0